pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localriskinsights"
version = "0.1.0"
description = "Local-first FastAPI + React transcription and risk insight workstation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "python-dotenv>=1.0,<2.0",
  "pyyaml>=6.0,<7.0",
  "uvicorn[standard]>=0.30,<1.0",
  "faster-whisper>=1.1,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
]
diarization = [
  "pyannote.audio>=3.3,<4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

